+++ python/aistore/sdk/batch/extractor/tar_stream_extractor.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2025, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2025-2026, NVIDIA CORPORATION. All rights reserved.
 #
 
 import tarfile
@@ -73,12 +73,12 @@ class TarStreamExtractor(ArchiveStreamExtractor):
 
                             # Get MossOut (from response or build from request)
                             moss_out = self._get_moss_out(
-                                index, len(content), moss_req, moss_resp
+                                index,
+                                len(content),
+                                moss_req,
+                                moss_resp,
+                                member_name=tarinfo.name,
                             )
-
-                            # Check for missing
-                            # TODO: FIXME: what to do with missing files?
-                            # moss_out.is_missing = tarinfo.name.startswith(GB_MISSING_FILES_DIR)
 
                             index += 1
                             yield moss_out, content